Move simpleflow.swf.process' Poller to a more explicit file
