+++ simpleflow/settings/logging_formatter.py
@@ -33,17 +33,17 @@ class SimpleflowFormatter(logging.Formatter):
     # Example of record dict:
     # {
     #     'threadName': 'MainThread',
-    #     'name': 'simpleflow.swf.process.actor',
+    #     'name': 'simpleflow.swf.process.poller',
     #     'thread': 140735241315072,
     #     'created': 1450436645.513802,
     #     'process': 84828,
     #     'processName': 'MainProcess',
     #     'args': (),
-    #     'module': 'actor',,
-    #     'filename': 'actor.py',
+    #     'module': 'poller',,
+    #     'filename': 'poller.py',
     #     'levelno': 20,
     #     'exc_text': None,
-    #     'pathname': '/path/to/simpleflow/simpleflow/swf/process/actor.py',
+    #     'pathname': '/path/to/simpleflow/simpleflow/swf/process/poller.py',
     #     'lineno': 76,
     #     'msg': 'starting <bound method ActivityPoller.start of <simpleflow.swf.process.worker.base.ActivityPoller>',
     #     'exc_info': None,
