+++ simpleflow/swf/process/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import absolute_import
 
-from .actor import Poller  # NOQA
+from .poller import Poller  # NOQA
 from .decider import Decider  # NOQA
