+++ simpleflow/swf/process/decider/base.py
@@ -8,7 +8,7 @@ import swf.format
 import swf.models.decision
 
 from simpleflow.process import Supervisor, with_state
-from simpleflow.swf.process.actor import Poller
+from simpleflow.swf.process import Poller
 
 
 logger = logging.getLogger(__name__)
